+++ src/extension/Remoting.fs
@@ -1,4 +1,4 @@
-module Remoting
+module OpenXmlExplorer.Remoting
 
 open Fable.Core
 open Fable.Import.Axios
@@ -27,6 +27,7 @@ let private getApiClient (serverHost) : IOpenXmlApi =
             |> toAsync
             
         stopApplication = fun () ->
+            Log.line $"Stopping API Server ..."
             axios.get(getRoute "stopApplication")
             |> toAsync
     }
@@ -34,7 +35,6 @@ let private getApiClient (serverHost) : IOpenXmlApi =
 open Node
 open Node.ChildProcess
 open Fable.Core.JsInterop
-open Fable.Import
 
 let private toStr = function
   | U2.Case2(x:Buffer.Buffer) ->
@@ -43,17 +43,16 @@ let private toStr = function
 
 let startServer port extensionPath =
     let cb (e:ExecError option) stdout' stderr' =
-      let channel = vscode.window.createOutputChannel "openxml"
-
-      channel.appendLine($"Out: %s{stdout' |> toStr}")
-      channel.appendLine($"Err: %s{stderr' |> toStr}")
-      if e.IsSome then
-          channel.appendLine($"ExecError: %s{e.Value.ToString()}")
-          channel.show()
+        //stdout' |> toStr |> fun s -> if s <> "" then Log.line s
+        stderr' |> toStr |> fun s -> if s <> "" then Log.line s
+        if e.IsSome then
+            Log.line($"ExecError: %s{e.Value.ToString()}")
+            Log.show()
 
     let host = $"http://0.0.0.0:%d{port}"
     let opts = createEmpty<ExecOptions>
     opts.cwd <- Some (extensionPath + "/bin")
     childProcess.exec ($"dotnet Server.dll %s{host}", opts, cb) |> ignore
+    Log.line $"API server started at %s{host}"
 
     getApiClient host